Fix 'check' command to return bad notes
If any notes link to a vertex that does not exist, the 'check' command
will list that note.

The validation will now have to take place within the Notebook instead
of being applied as a filter as first hoped. This is due to the graph
not being built with invaild notes due to them refering to missing
verticies.