Stop URL anchors from being treated as a tag
A tag must now be preceded by whitespace only for it to be treated as
a tag. This avoids other notations using '#' to accidently be
matcheed as a notebook tag.

For some reason we were supporting tags inside quotes. I don't appear
to have used this in my notebook so dropping support for this to fit
with the stricter tag matching.