Start creating 'context' function
The aim of this is to make it possible to extract a related paragraph or
sentence from some source text that contains a given term or phrase.

This could then be used when searching for related things in the
notebook to output a contextual match within the body of the note rather
than just outputting note titles.