Fix no results when matched more tags than provided
Due to fuzzy tag matching, the matched tags produced a larger set of
tags to search by than the original list we provided leading to
unrelated tags being matched and included in our 'AND' search.

Whilst we want fuzzy tag matching for ease of use, we don't want
it to pull in unrelated tags. Resolve this by comparing the
Levenshtein distance between our provided tag names and matched tags.