Support tags with multiple leading hashes
Allowing support for several hashes at the start of the tag name
will allow them to be used for structure tags. A regular tag will allow
notes to be retrieved that refer to thing, but using multiple hashes
for the tag allows seacrhing for the top-level note on that concept.