Fix off by one error when getting matched line number
The note parser extracts the note starting line number which includes
the note title. When getting the matched line number on gathering
context it will recount the title line again as that's its line 1.

The Note 'Content' property will now always include the title line
becauese if a note has a markdown title it would be excluded from the
content but if there isn't a markdown title, the title of the note will
be the first line of the main text which means when calculating the line
number we won't know if the title was included or not, so make the
behaviour consistent.