Fix incompatibility with being given 'fzf' query strings
'fzf' appears to pass a single quoted query string to "bind" commands
which 'zb tag' will treat as a single tag name, but multiple tags won't
match because tags don't contain spaces.

If we're given a single argument that has spaces in it, break the string
up by splitting on spaces and treating each term as an individual tag as
this is more likely what the user meant.