Add git integration for recent note sorting
'zb' doesn't care what vertex IDs are so long as they're a number and
recent notes assumes that a larger vertex ID means it is a newer
reference and so will rank them higher.

A digital Zettelkasten (or knowledge garden) tend to use timestamps for
IDs because they're easy to create manually or via a script. Making an
assumption that the IDs are timestamps and accessing notes within a
Git repository we can make it easier to revisit notes that have been
edited and committed to since creation. Edited notes may temporarily be
more relevant so try to include commit date as a factor when sorting
the knowledgebase to return the recent note order.