Allow writing local and global configuration
Can write the entire current state of the config to the intended
destination using 'config --global --save' but 'config --global' won't
print out the global configuration because we've already loaded the
configuration context based on where 'zb' has been called.

It might be better to support the idea of global and local partial
configurations that build the complete configuration allowing us to
globally set our editor of choice but having notebook directories
specific to the notebook folders.