Make it possible to unset configuration values
Viper doesn't make it easy to delete a config key that was already set
so we can blank it but that won't mean that 'zb config' will return
to its original state of showing the default values for config keys
because as far as Viper is concerned the blank keys have a configured
value.

This edge case can be resolved by manually editing the config flie and
deleting the key after it has been unset.