Remove local config scanning
A Zettelkasten / Digital garden should tend to be 1 thing. The
likelihood of having several that we want to search independent of one
another hasn't come up in my use of my notebook so far.

If we remove the concept of automatically having a "local" config for
notebook(s) we can make managing the config with different options
easier because we're no longer swapping between different config files.

We still have the option to manually provide the location of the config
which will still allow overriding this, but the user is explictly saying
switch to this context so there's no ambiguity of what config we should
use.