Fix failure to set config option that was not already in config file
When attempting to set a config option we call 'viper.InConfig'
expecting it to check if the provided option is a valid option for our
config, but it's actually saying is it already set in the config file.

Crudely create a routine that knows what config options we expect to be
settable in our config file to workaround the problem. There might be a
better way but this works for now.