Fix duplicate directories being added to config on first init
'zb' has tried to do something useful if it didn't have any notebook
directories configured yet where it could assume the current directory
or 'notes' directory if one exists.

When creating our initial 'init' we extend the current config, which
would initially assume your current directory leading to an extra path
being written to the config file when the file wasn't there.

It makes more sense now to leave 'zb' without any notebook directories
when first called so that the user can configure the notebooks as
required and will get an initial experience that is easier to
understand.