Do not output notes as JSON as default behaviour
Needing to read the notebook as JSON is not a feature that has been used
often so it shouldn't be the default behaviour. Now that 'zb' will not
try to assume the notebook directory if it doesn't have it configured it
makes sense to output help content by default and move JSON output under
its own subcommand.