Move loading of files out of Notebook
A Notebook can now just be a collection of files that should be built
into a graph that can be queried. Hopefully this makes this easier
to share as a library because the Notebook is longer concerned with
how the files are sourced so long a it has a collection of files to
read.