Make Note available in top-level 'zb' package
To allow an external project to make use of the notebook it would
be useful if the external project has easy access to the Note struct
that zb parses files into. To make this possible the 'main' package
needs to be moved under 'cmd/zb' for the command-line utility so that
it does not collide with the 'zb' package namespace.

Add a Makefile to make it easier to build the executable given it
has now been moved from the top-level directory and 'go build' will
no longer be enough to bulid the project as expected.